requirements.txt
customtkinter
keyboard
pystray
pillow